Add section on NEMO-3.1 & CONCEPTS-110 reference repos.
Includes description of the construction of the NEMO-3.1 repo.
That repo was the original NEMO-code repo before we decided that porting
CONCEPTS to NEMO>=3.4 was somebody else's job.
The description of the repo construction is mostly a copy of the Initialization
to NEMO-3.1 section from `nemo-code.rst` that will soon disappear from that
file.