Docs maintenance (#56)
* Update ReadTheDocs build config

Upgraded the OS to Ubuntu 24.04 and updated the Python toolchain to
mambaforge-23.11 in the .readthedocs.yaml file. This ensures
compatibility with the latest environment settings available on
readthedocs.

* Update Sphinx & optional dependencies version pin

Updated Sphinx and optional dependencies packages pinned versions in
`environment-rtd.yaml` in preparation for Python 3.13 environment.

* Drop Sphinx source_suffix config

The deprecated list of strings appears to have been removed in
Sphinx=8. The default `source_suffix = {'.rst': 'restructuredtext'}` is
appropriate for this project.

* Update redirected links found by linkcheck

* Update "Existing storm surge models" section

Replace description and link for discontinued Storm Surge BC site
with description and link to ECCC Coastal Flooding Prediction and
Alerting Program.

* Add pre-commit hooks

* Update broken CHS water level data links